riverlea - queue individual files with bundler rather than @import
