+++ ext/riverlea/core/css/_base.css
@@ -1,10 +1,4 @@
 /******************
-   VARIABLES
-*******************/
-
-@import url(_variables.css); /* default variables */
-
-/******************
    RESETS
 *******************/
