+++ ext/riverlea/core/css/civicrm.css
@@ -1,4 +1,1 @@
-@import url(_base.css); /* resets and civi wide styles */
-@import url(_cms.css); /* CMS-specific hacks/workarounds */
-@import url(_core.css); /* components and utility classes */
-@import url(_fixes.css); /* deprecated-Civi handling, hacks/workarounds. one day this will go. */
+/* This file is intentionally blank. It is currently overloaded by Thames/Walbrook to add custom rules. */
