+++ ext/riverlea/streams/walbrook/css/civicrm.css
@@ -1,8 +1,3 @@
-@import url(../../../core/css/_base.css); /* resets and civi wide styles */
-@import url(../../../core/css/_cms.css); /* CMS-specific hacks/workarounds */
-@import url(../../../core/css/_core.css); /* components and utility classes */
-@import url(../../../core/css/_fixes.css); /* deprecated-Civi handling, hacks/workarounds. one day this will go. */
-
 /* Font */
 
 @font-face {
